auto-mailer: for root branches, always stop iterating on sent commits
It should have been obvious that the merge-base of the new version of
the ref and the old version of the ref does not necessarily
blocks the iteration of the tree. So, for the case of 'root refs' (such
as master) being pushed, we should stop at the first commit wr see
that was already sent and not continue until we reach the bottom.