Pick leading branch according to sorted refs
The branch name in the subject line was arbitrary the first one
in the lexical sort of contained branch, irrespective of the branch
from which the commit was detected when iterated the sorted ref
list.

This change also removes the ignore branch names from the list
of branches in each commit, to make stuff less noisy.